Check that feature image exists before outputting
If we pass an image that doesn't exists to the rails image_tag helper,
the rails app will break. Therefore, we check to see that the image
exists before outputting the image. This is a bit slow since we have to
actually ping our S3/Dropbox. There might be a better solution.